requirements.txt
absl-py==2.1.0
altgraph==0.17.4
blinker==1.7.0
boto3==1.34.89
botocore==1.34.89
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cx_Freeze==7.0.0
cx_Logging==3.2.0
DateTime==5.5
docutils==0.21.1
easygui==0.98.3
et-xmlfile==1.1.0
Flask==3.0.3
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
gunicorn==23.0.0
idna==3.7
immutabledict==4.2.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
lief==0.14.1
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
ortools==9.9.3963
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
playsound==1.3.0
plyer==2.1.0
protobuf==5.26.1
psutil==5.9.8
pycparser==2.22
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtdeploy==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
qrcode==7.4.2
requests==2.31.0
s3transfer==0.10.1
scipy==1.13.0
six==1.16.0
tk==0.1.0
toml==0.10.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
watchdog==4.0.0
Werkzeug==3.0.2
zipp==3.18.1
zope.interface==6.3
